obsoleted OCLValidationDelegate.constraintNameMap/invariantOperationMap
Parse invariants into Constraint elements. This required change of one test case
where "invalid" is parsed and an error is expected. The error message produced when
parsing an invariant into a Constraint is slightly different from the error message
produced when parsing a query and then checking it in the OCLValidationDelegate
for Boolean result type.