ClassChooserPage added to the wizard
Upon selecting an existing .ecore file either from workspace or from a
registered package one will be able to choose one class from the metamodel
which will be entered into the .tcs file.
Biggest TODO: Find a way to give the SyntaxBuilder an appropriate reference
to the metamodel and make sure the .tcs file is correct.