big projectwizard update
When creating a project with a metamodel project the Parser and Lexer classes
can be generated by repeatedly building the project. The folder generated/
generated has to be refreshed in order to make the Builder "see" the .g file.
The projectwizard.test now runs with the new metamodelproject.
It tests the .tcs file that the wizard creates.
Still todo: -Missing classes in the src folder of the language project
            -Editor throws error due to inconsistent hierarchy
            -The ClassViewerPage needs a Listener
	    -Loading an existing Metamodel from workspace or registered
             packages is not fully functional yet
	    -comments and a readme for changing templates due to new versions