Fixed class loading for ParserGenerationTestHelper; fixed OCLs in TCS.tcs
ParserGenerationTextHelper.getparserFactoryClass now also uses a ClassLookup
to be able to load classes in the scope of a particular (test) bundle's
class loader scope.
Added first test class (not green yet) TestIncrementalParser to
referenceresolving.tests in order to enable text blocks-based parsing and
testing.
Added an opposite reference in TCS.ecore for the connection between a
ForeachPropertyInit and its PredicateSemantic element which is needed by
some OCLs in TCS.tcs.
Adjusted OCL in TCS.tcs which made reference to renamed references and
therefore broke. Added a cast of an eType from EClassifier to EClass to make
OCL expression compile.