Partly fix the delayed reference resolving fixpoint iteration.
The fix is only partial as it fails to account for the potentially
required reevaluation of foreach predicates. The problem is that those
tirgger the creation of additional delayed references.

A potential solution is to implement appropriate equals-methods on the
refactoed DelayeReference classes, or to entirely rely on the impact
analysis.