find context template and from there path to elements created by those containing #context in OCL
The code in AbstractFurcasOCLBasedModelUpdater navigates from the context element to the
text block documenting its creation and from there descends the text blocks tree until it
finds text blocks documenting the execution of the template using the #context in an OCL
expression with the right tag.

Still missing is the filtering out of those template executions where the actual sequence
element containing the OCL expression did not fire, e.g., because a different alternative
was chosen.