In tear-down deregister trigger manager to avoid conflicts during standard parsing
Furthermore, I relaxed the testAddArticleAndExpectRevenueLedgerCreation test slightly.
It doesn't check the contents of the RevenueLedger object created (which aren't set yet).
The reason they aren't set yet is that during the RevenueLedger construction the OCL IA
isn't triggered because the updaters use #foreach/#context and hence their context type
differs from the type of the element to update.