fixed bug introduced in previous commit regarding getParent() on null textblocks
The TbParsingUtil.getParent(TextBlock) operation returned null for a null argument.
I first refactored into a plain call to getParent() on the text block. That was
slighly non-semantic-preserving as it causes a NullPointerException for null
TextBlock arguments. The following inlining of the method body was semantic-preserving
but multiplied the potential problem to all previous calls to the method. I fixed
those areas where test cases failed.