Migrate the remaining pretty printer tests.
Some of them are still failing because the current TCS.tcs is flawed. It
uses propertyInits to update the keyword list of the ConcreteSyntax
depending on the current content of this list. This is a violation of
the protocol that forces the pretty printer to only chose templates
for which the values of proprties are equal to the result of their
corresponding propertyInit expression.