trying to reduce memory consumption of OCL impact analysis benchmark
In this case, I've tried to avoid redundant size computations on resources
which each traversed all elements in the respective resource. This seems to
have created abundant iterator objects. Instead, I'm now using a specific
ResourceImpl subclass which caches the size and copies the size from one
clone to the next.