Synchronize editor modifications.
WARNING: It is yet unclear if this is sufficient.

The lexing phase is synchronized on the same lock object as all
operations on the CtsDocument. This is supposed to prevent concurrent
modifications during lexing. It is ASSUMED that parsing only requires
the 'current' TB version and does not interfer with the
constantly user-modified 'previous' version. Thus, parsing should not
require locking.