fixed issue regarding parsing foreach propinits without textblocks support
So far we were trying to remember the Template object used to produce the foreach
target element in the ForEachContext object in the textblocks model. This and the
comparison of Template objects is not ideal, particularly as when parsing
without textblocks support the parser doesn't attach the property inits
to the foreach DelayedReference, making it impossible to find out the Template
in the first place. Instead, we're now using the parser production rule name.