fixed foreach update for changes within the same when-clause
Back to storing the Template in the ForEachContext, only this time more carefully
so that when not in text blocks creation mode then no exception occurs. The template
is then used to compare the owner of the foreach property init (a template) with the
template used to produce the element within which a property would have to be updated.