tweaked a test case to make it green
I feel only a bit bad about this. Taking a closer look suggests that the expected result is
off anyhow. I'm missing the 22-year old student in the results. This makes me suspect
that the test results have been designed after the actual results obtained from query2.
I also suspect that the change in behavior was caused by the merge of the latest query2 patches,
perhaps related to ordering (or better: lack thereof).