pre-applying the changes of https://bugs.eclipse.org/bugs/show_bug.cgi?id=339052
Rolling back the changes of 9a5bec7ec047b9f1a9127aac0c6f737375301863
which will not make it into the standard OCL implementation for unprecise
semantics of package lookup. This commit allows clients to use a specific
EPackage.Registry and put ambiguous packages into the registry under
distinct URIs where the URIs have the form of a legal OCL package name
which is either a simple name or a double colon-separated name.