pre-applying the changes of https://bugs.eclipse.org/bugs/show_bug.cgi?id=339052
Rolling back the changes of 63be6c146ff70926a3986bbe83eb7c75e95cdcf3
which will not make it into the standard OCL implementation for unprecise
semantics of package lookup. This commit allows clients to use a specific
EPackage.Registry and put ambiguous packages into the registry under
distinct URIs where the URIs have the form of a legal OCL package name
which is either a simple name or a double colon-separated name.