Fix setting of the correspondingModelElements.
Due to an interface/information passing issue the pretty printer does
not know which element shall be added to the referencedElements of
tokens. It depends on the parser to do this upon reparsing of all
textblocks.

This is a real problem. A proper solution seems to be adapting the
pretty printer to create a trace which conforms to the interface of
the ParserTextBlocksHandler.