Adjusted the semantics of the EventManagerTableBased implementation to the matchesFor methods
This uses Bags instead of Sets internally, counting the number of matches each registration
requires per table. This also covers contradictory conditions (A and not A) as well as
multiple registrations for the same table in the same AndFilter. Added corresponding
test cases. Moved more stuff away from the EventManager's API package.