added test case and provided fix for changing when-clause
When a foreach result changed such that first no when-clause applies and
later some when-clause does apply, no element was added. Works now.
More potential bugs identified for which test cases and fixes have to follow;
particularly, don't expect a change from some when-clause match to no
when-clause match to work properly so far.