made test stubs implement InternalEObject including a valid eClass() impl for use in getID(...)
The tests failed because for property inits we now ask their URI which performs a cast to
InternalEObject and asks the object's eClass.
Additionally, a few tests needed to be adjusted regarding their expected outcome because
they tested on the particular setOclRef call signature generated which now includes the
additional parameter.