Lazy load EcorePackageLoadListener on the first notification.
This is supposed to prevent cyclic bundle / class loading: The
SyntaxRegistry forced the EMFBundleListener to load, which in return
forced all registed EcorePackageLoadListeners to load. The
SyntaxRegistry is included in the latter but not completely loaded (as
it still waits for the initialization of the EMFBundleListener) and
thus triggers a framework exception.