Fix the lookup for templates terminating an operated template.
For some reason it is not allowed to add abstract templates to an
inheritance hierarchy which is started with an operated template.
This fix makes it possible to 'skip' over these non-existing in-between
templates when trying to find a suitable terminating template.

The comment attached to the code indicates that it was always meant to
be done this way. Seems like the code broke when getTemplates was
adapted return a whole collection instead of a single template.