Fix operator associativity bug due to missuse of enums and conditionals.
This is wrong: (associativity = 'left' : ? "," associativity) "{"...

Using this form of a conditional sets the associativity to 'left' when
the lookahead is "{" . "When the lookahead is "," the associatity
following the "," is parsed and set. So far so good.

However, immediately after that the associativity is overwritten
by 'right', which is imposed by the else-part of the conditional.
The actually parsed associativity is thus ignored.