added autoResolve property to dslEngineering.properties
the property controls if referenced metamodels are resolved
automatically or not. While Stephan needed automatic resolvability of
referencing metamodels that approach confliced when a furcas project
that uses the furcas metamodel was created. autoResolve's default value
is <false> - thus referenced metamodels won't be resolved automatically
unless the property is set to true.