fixed two more issues in incremental parser
now two more tests are running green.
Reason was: When the EOS was reached and the last remaining tokens were touched, only the block being the common ancestor of those last tokens was reparsed. In fact, as there may have been more tokens that have been removed up to EOS we have to re-parse the root block instead.