fixed IllegalArgumentExcetion while trying to save a textual model
The comment above method getWorkspaceResources() indicates that the
intention of this method is to filter plugin resources that cannot be
changed an thus don't need synchronization. However the ArrayList is
never filled because of a missing add-method call. I've added this
method call and now saving a textual model is possible again.