Implement and test identifier recovery for broken references.
It may happen that a referenced model element is gone. If an
existing textblock still contains the identifier that was
used to originally reference this model element, the newly printed
textblock will re-use the identifier. That way, the reference can be re-
established at a later point in time (e.g., by the parser).