Prevent crash on empty leading sub-TextBlocks.
During the recursive TextBlock instatiation and token move operation, we
may have to navigate to back to our parent block. At this time this
block is not yet completely created and may contain empty leading
subblocks. We must be able to deal with this inconsistent state.

We do now skip over these empty blocks. The parser will no longer crash.

Unfortunately, I am not absolutely sure that this is the correct
fix in all cases. It may potentially happen that we do compute
the wrong original version. Further investigations / reviews
are therefore needed.