Remove risky TB reuse during merge.
The incremental parser will now consider less TextBlocks to be equal.
This fixes some problems, but we may re-use fewer elements than we used
to! There might be a better way to fix the problem described below.

We might have considered TextBlocks to be equal which have different number
of tokens. Those tokens then ended up as a duplicate in the re-used TB,
resulting in an invalid TB model.

We only check for re-usable tokens on the current sub-node level (the level
where the lexer created these new tokens) and might then confuse tokens to
be equal, though these do actually belong to different sub-blocks in a
correctly nested TB tree.

There seems to be no obvious way to remove these outdated, remaining tokens.

This fixes the problems observed in TestParsingScenarios#
testReparsingWithoutModiication*();