Only unset properties which are completely unused.
There may be more than one reference to a single property within
a textblock (e.g.,the auto-parenthesis within operator templates
have the same sequence element as the property itself). We
therefore need to find out if a property is still 'referenced' by any other
token in the re-used TextBlock.
Only if it is no longer referenced from anywhere, we are free
to unset it on the corresponding modelelement.