Continue lexing until token types converge.
When deciding whether we can stop lexing or if we have to keep going,
we checked if we have two off-channel tokens at hand. Having two tokens
of this same time implied that our re-lexed stream converged with the
unmodified tokens.

This checked failed when we commented a line. We stopped lexing too early
and did not properly lex the last token at the end of the commented
line.

This commit fixes this by checking the more fine grained token type.