Use a SyntaxRegistry per editor (no more singleton)
Syntax sharing does not work. For a syntax to be shared, either the
ResourceSet needs to be shared or the syntax has to be loaded using the
delegation mechanism of a ResourceSet during loading.

The problem is that once we have such a deferred loading in place, we
can no longer resolve cross-references as those would need to be
resolved cross resource sets (the editor one and the one where we keep
the syntax).